Display Artist additional profile information

+++ database/migrations/2017_06_17_091944_add_category_column_to_artist_table.php
@@ -27,9 +27,12 @@ class AddCategoryColumnToArtistTable extends Migration
     public function down()
     {
         
-        Schema::table('artists', function (Blueprint $table) {
+        //Remove Artist Categroy foreign key from Artist table
+        if (Schema::hasColumn('artists', 'artist_category_id')) {
+                 Schema::table('artists', function (Blueprint $table) {
            $table->dropColumn('artist_category_id');
             
         });
+        }
     }
 }
